Backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
fastapi-cors
gspread
oauth2client
pytz